requirements.txt
redis==4.5.5
SQLAlchemy==2.0.13 
python-dotenv==1.0.0 
psycopg2-binary
loguru